Refactored offline functions
