+++ domains/DataCollection/Forms/PatientID/index.js
@@ -7,13 +7,15 @@ import {
 import { Text, Button } from 'react-native-paper';
 import { Formik } from 'formik';
 // import * as yup from 'yup';
-import * as Network from 'expo-network';
 import { postObjectsToClass } from '../../../../services/parse/crud';
 import PaperInputPicker from '../../../../components/PaperInputPicker';
 import configArray from './config';
 import {
-  storeData, getData, getAllData, deleteData
+  storeData
 } from '../../../../modules/async-storage';
+import { checkOnlineStatus } from '../../../../modules/offline'
+import { backgroundPostPatient } from './utils';
+import { generateRandomID } from '../../../../modules/utils';
 
 // const validationSchema = yup.object().shape({
 //   fname: yup
@@ -30,49 +32,15 @@ const PatientIDForm = ({ navigation }) => {
   // similar to componentDidMount and componenetWillUnmount
   // runs every 10 seconds in the background to get all Async Data
   useEffect(() => {
-    const interval = setInterval(() => {
-      checkOnlineStatus()
-        .then((isConnected) => {
-          if (isConnected) {
-            getAllData().then((allAsyncData) => {
-              // contains all the available keys
-              const allKeys = allAsyncData.map((a) => a[0]);
-              allKeys.forEach((item) => {
-                if (item.includes('PatientID-')) {
-                  getData(item)
-                    .then((postParams) => {
-                      postObjectsToClass(postParams)
-                        .then(() => {
-                          deleteData(item);
-                          toRoot();
-                        }, () => {
-                        });
-                    });
-                }
-              });
-            });
-          }
-        });
-    }, 10000);
-    return () => {
-      clearInterval(interval);
-    };
+    backgroundPostPatient().then(() => {
+      toRoot();
+    })
   }, []);
 
   const toRoot = () => {
     navigation.navigate('Root');
   };
 
-  const generateRandomID = () => Math.random().toString(20).substr(2, 12);
-
-  // checks whether user is connected to internet, return true if connected, false otherwise
-  // maybe on componentDidMount calling this function and then create another function to upload
-  // the objects and delete them afterwards. NEEDs to BE FLESHED OUT
-  async function checkOnlineStatus() {
-    const status = await Network.getNetworkStateAsync();
-    const { isConnected } = status;
-    return isConnected;
-  }
   const [inputs, setInputs] = useState({});
 
   useEffect(() => {
