+++ modules/utils/index.js
@@ -1,0 +1,3 @@
+const generateRandomID = () => Math.random().toString(20).substr(2, 12);
+
+export { generateRandomID };
